refactory from square to circles done
made it so that all squares are circles instead, mainly to showcase the bouncing of the wall mechanics and throw velocity better.

collision detection when interfering with other objects seems to be good but need to double check when ive corrected the surface sinking thats happening right now.
to implement:
- circles are surpassing to an extent beyond the floor level.
- add space gravity toggle to showcase the collision better.
- [IMPORTANT] implement square bar to showcase friction